% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_article.R
\name{datapoint}
\alias{datapoint}
\title{CFPB Data Point Format}
\usage{
datapoint(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a CFPB Office of Research Data Point.  This will generate
a PDF file that is suitable for publication.
}
