% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{dp_kable}
\alias{dp_kable}
\title{Data Point Kable}
\usage{
dp_kable(x, format = NULL, digits = getOption("digits"), row.names = NA,
  align = NULL, caption = NULL, format.args = list(), escape = TRUE,
  booktabs = TRUE, latex_options, linesep = "", ...)
}
\description{
This function overrides the knitr::kable function. For LaTeX documents,
it adds code to color the top row of the resulting tables. For other
document types, this function is equivalent to knitr::kable.
}
