% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{dp_palette}
\alias{dp_palette}
\title{Data Point Palette}
\usage{
dp_palette(x)
}
\arguments{
\item{x}{A vector of numbers between 1 and 8}
}
\value{
A hex representation of the appropriate color
}
\description{
This function returns the color palette that we used for Data Points.  It contains
colors from the CFPB palette, alternating between 'warm' and 'cool' colors as
the Design team recommended.
}
