% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{dp_theme}
\alias{dp_theme}
\title{Data Point Theme}
\usage{
dp_theme(base_family = "Helvetica", base_size = 11,
  light_family = "Helvetica-Narrow")
}
\arguments{
\item{base_family}{Font family to use (default = 'Helvetica')}

\item{base_size}{Size of the base font to use (default = 11)}

\item{light_family}{Font family to use where light text is required}
}
\description{
This is the ggplot2 theme that I used to create the graphs in the Data Point.
}
