% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoint.R
\name{datapoint}
\alias{datapoint}
\title{CFPB Data Point Format}
\usage{
datapoint(qCCT = F, ...)
}
\arguments{
\item{...}{Arguments to \code{bookdown::pdf_book}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown::render]{render}}
}
\description{
Format for creating a CFPB Office of Research Data Point.  This will generate
a PDF file that is suitable for publication.
}
